keywords hits on gene xml
